+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log
 
-All notable changes to the "p5.js Project Generator" extension will be documented in this file.
+All notable changes to the "p5.js 2.x Project Generator" extension will be documented in this file.
+
+## [1.0.2]
+
+- Added IntelliSense and autocomplete support for most p5.js functions
+- Added hover documentation for p5.js functions
+- Added signature help with parameter hints
+- Automatic detection of p5.js project files for autocomplete activation
+- Removed deprecated `preload()` function from autocomplete
+
+## [1.0.1]
+
+- Updated display name to "p5.js 2.x Project Generator"
 
 ## [1.0.0]